Servicio de Hist. Oper. y ABM varios

+++ Models/Products.cs
@@ -8,7 +8,7 @@ namespace AspNetCoreBank.Models
         public int Id { get; set; }
         public int ClientId {get; set;}
         public decimal current_balance { get; set; }
-        public decimal status { get; set; }
+        public int status { get; set; }
         public DateTime Opening_date { get; set; }
         public DateTime? Closing_date { get; set; }
         public virtual Client Client { get; set; }
